+++ src/parser/utils/webpackTemplates.js
